requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
distro==1.9.0
Flask==3.1.0
flatbuffers==25.1.24
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
openai==1.61.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
rapidocr-onnxruntime==1.2.3
requests==2.32.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
azure-cognitiveservices-speech
aiohttp
requests
elevenlabs